inference-service/requirements.txt
numpy
fastapi
uvicorn[standard]
pydantic
scikit-learn
xgboost
pandas
numpy
debugpy
grpcio
grpcio-tools